solved sorted array
---
* The iterators should be used with == or != boolean comparison
operators

* intial method of trying with left_itr <= right_itr wasn't working.
TODO check why the pointed values were getting compared instead of the
the iterators themselves
* since the array is sorted and the values are squared, the leftmost
value will be the smallest and the rightmost value will be the largest
when compared together with the sign. If we are squaring, then the
leftmost value can also be the largest. So created a final sorted array
of sorted squared values, we start inserting from the rightmost end of
new array.