algo2
---
using inplace reversal of the string by checking the space and
reversing the string. We use a special if condition on the string index
(i+1) > string.size() get the last word in the string.
Also there is a sentinel last_word_index which is assigned the begining
of next word each time we detect a space using i+1. ith index will be
the space.