check for substring using hash_map
--
this approach uses a sliding window apporach wherein, we first create
two hash_map which has letters as keys and their corresponding freq.
For the second string, we only use the substring equal in length
to the first string, and update the map by sliding the window across
at each iteration, we decrement the count of the outgoing letter (and
erase it also if the count decreases to 0, for comparing hash map)
and increment the incoming letter frequency.