added code for group group_anagram
--
The idea behind this is to create a hash from the words. Since
anagram will have same character count, we can create a key based on
a fixed length 26 of letters as a character array, and use it as a key
after it is converted to a string ( C++ can convert string to hash)
by itself. For using a character array we need to define a hash function