added two sum problem
---
Since this question requires to solve the problem using Constant Space,
We can't use hash map and search for the lookup element
(target - nums[i]) stored in the hash map.
We can use the two pointer approach to solve this problem, where one
pointer moves from the start and other from the end, and Since
the array is sorted we can increment the pointer from start if the
sum of ith and jth element is less than target and decrement the
right pointer if the sum is greater than target.